move logic to useGapsLsit hook
